+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.0 - 2025-04-16
+
+- Added `prefetch()` function
+
 ## 0.1.0 - 2025-04-16
 
 - Added `ShadowBridge` class